# docs/swagger.yaml
basePath: /
definitions:
...
  dto.DefaultDisappearingRequest:
    properties:
      timer:
        description: 24h, 7d, 90d, or off
        example: 24h
        type: string
    required:
    - timer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: session not found
        type: string
    type: object
...
  dto.PrivacySettingsResponse:
    properties:
      settings: {}
    type: object
  dto.ProfileInfoResponse:
    properties:
      profile: {}
    type: object
...
  dto.SetNameResponse:
    properties:
      name:
        example: John Doe
        type: string
    type: object
  dto.SetPictureResponse:
    properties:
      pictureId:
        example: abc123
        type: string
    type: object
...
  dto.SetPrivacyRequest:
    properties:
      setting:
        description: last_seen, online, profile, status, read_receipts, group_add,
          call_add
        example: last_seen
        type: string
      value:
        description: all, contacts, contact_blacklist, match_last_seen, known, none
        example: contacts
        type: string
    required:
    - setting
    - value
    type: object
  dto.SetProfilePictureRequest:
    properties:
      image:
        example: base64_encoded_image
        type: string
    required:
    - image
    type: object
  dto.SetPushNameRequest:
    properties:
      name:
        example: John Doe
        type: string
    required:
    - name
    type: object
  dto.SetStatusRequest:
    properties:
      status:
        example: Hey there! I'm using ZPWoot
        type: string
    required:
    - status
    type: object
  dto.SetStatusResponse:
    properties:
      status:
        example: My new status
        type: string
    type: object
...
info:
  contact: {}
  description: WhatsApp API with Chatwoot Integration
  title: ZPWoot WhatsApp API
  version: 0.1.0
paths:
...
  /sessions/{sessionId}/profile:
    get:
      consumes:
      - application/json
      description: Get own WhatsApp profile information
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get own profile info
      tags:
      - profile
  /sessions/{sessionId}/profile/disappearing:
    patch:
      consumes:
      - application/json
      description: Set the default disappearing messages timer for new chats
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Timer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DefaultDisappearingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set default disappearing timer
      tags:
      - profile
  /sessions/{sessionId}/profile/name:
    patch:
      consumes:
      - application/json
      description: Set WhatsApp display name (push name)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Name data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetPushNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SetNameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set display name
      tags:
      - profile
  /sessions/{sessionId}/profile/picture:
    delete:
      consumes:
      - application/json
      description: Remove WhatsApp profile picture
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete profile picture
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Set WhatsApp profile picture
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Image data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetProfilePictureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SetPictureResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set profile picture
      tags:
      - profile
  /sessions/{sessionId}/profile/privacy:
    get:
      consumes:
      - application/json
      description: Get WhatsApp privacy settings
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PrivacySettingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get privacy settings
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Set WhatsApp privacy settings
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Privacy settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetPrivacyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set privacy settings
      tags:
      - profile
  /sessions/{sessionId}/profile/status:
    patch:
      consumes:
      - application/json
      description: Set WhatsApp status/about message
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Status data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SetStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set status message
      tags:
      - profile
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"
...